# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMONTA BENCANA API Documentation",
        "description": "Sistem Informasi Manajemen Bencana Terpadu - RESTful API untuk pengelolaan data bencana, laporan, monitoring, dan integrasi dengan BMKG.",
        "contact": {
            "email": "support@simonta-bencana.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        }
    ],
    "paths": {
        "/api/admin/system-monitoring": {
            "get": {
                "tags": [
                    "Admin Management"
                ],
                "summary": "System Monitoring (Admin Only)",
                "description": "Endpoint untuk monitoring sistem (hanya untuk Admin).",
                "operationId": "adminSystemMonitoring",
                "responses": {
                    "200": {
                        "description": "System monitoring data retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Admin system monitoring"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT authentication token. Dapatkan token dengan melakukan login ke endpoint /api/auth/login menggunakan username \"admintest\" dan password \"123456\"."
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}